# docs/swagger.yaml
basePath: /
definitions:
...
  purchases.PurchaseDetail:
    properties:
      amount:
        example: 155
        type: number
      cost:
        example: 15.5
        type: number
      id:
        example: 123e4567-e89b-12d3-a456-426614174002
        type: string
      item_code:
        example: ITEM-001
        type: string
      item_description:
        example: Sample Item Description
        type: string
      item_id:
        example: 123e4567-e89b-12d3-a456-426614174001
        type: string
      purchase_header_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      quantity:
        example: 10
        type: integer
    type: object
  purchases.PurchaseDetailRequest:
    properties:
      amount:
        example: 155
        minimum: 0
        type: number
      cost:
        example: 15.5
        minimum: 0
        type: number
      item_id:
        example: 123e4567-e89b-12d3-a456-426614174001
        type: string
      purchase_header_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      quantity:
        example: 10
        minimum: 1
        type: integer
    required:
    - amount
    - cost
    - item_id
    - purchase_header_id
    - quantity
    type: object
  purchases.PurchaseHeader:
    properties:
      code:
        example: PH-001
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
  purchases.PurchaseHeaderRequest:
    properties:
      code:
        example: PH-001
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    required:
    - code
    - created_at
    type: object
...
host: localhost:8125
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API per gestionar compres, ventes i stocks bàsica
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FRDY API
  version: "1.0"
paths:
...
  /purchases/details:
    post:
      consumes:
      - application/json
      description: Creates a new purchase detail with the provided information
      parameters:
      - description: Purchase detail data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/purchases.PurchaseDetailRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Purchase detail created successfully
          schema:
            $ref: '#/definitions/purchases.PurchaseDetail'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new purchase detail
      tags:
      - purchases
  /purchases/details/{header_id}:
    get:
      consumes:
      - application/json
      description: Retrieves all purchase details for a specific purchase header
      parameters:
      - description: Purchase header ID
        in: path
        name: header_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of purchase details
          schema:
            items:
              $ref: '#/definitions/purchases.PurchaseDetail'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get purchase details by header ID
      tags:
      - purchases
  /purchases/details/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a purchase detail by its ID
      parameters:
      - description: Purchase detail ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Purchase detail deleted successfully
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a purchase detail by ID
      tags:
      - purchases
    put:
      consumes:
      - application/json
      description: Updates an existing purchase detail with the provided information
      parameters:
      - description: Purchase detail ID
        in: path
        name: id
        required: true
        type: string
      - description: Purchase detail data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/purchases.PurchaseDetailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Purchase detail updated successfully
          schema:
            $ref: '#/definitions/purchases.PurchaseDetail'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a purchase detail
      tags:
      - purchases
  /purchases/headers:
    get:
      consumes:
      - application/json
      description: Retrieves all purchase headers
      produces:
      - application/json
      responses:
        "200":
          description: List of purchase headers
          schema:
            items:
              $ref: '#/definitions/purchases.PurchaseHeader'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all purchase headers
      tags:
      - purchases
    post:
      consumes:
      - application/json
      description: Creates a new purchase header with the provided information
      parameters:
      - description: Purchase header data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/purchases.PurchaseHeaderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Purchase header created successfully
          schema:
            $ref: '#/definitions/purchases.PurchaseHeader'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new purchase header
      tags:
      - purchases
  /purchases/headers/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a purchase header by its ID
      parameters:
      - description: Purchase header ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Purchase header deleted successfully
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a purchase header by ID
      tags:
      - purchases
    get:
      consumes:
      - application/json
      description: Retrieves a purchase header by its ID
      parameters:
      - description: Purchase header ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Purchase header found
          schema:
            $ref: '#/definitions/purchases.PurchaseHeader'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a purchase header by ID
      tags:
      - purchases
    put:
      consumes:
      - application/json
      description: Updates an existing purchase header with the provided information
      parameters:
      - description: Purchase header ID
        in: path
        name: id
        required: true
        type: string
      - description: Purchase header data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/purchases.PurchaseHeaderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Purchase header updated successfully
          schema:
            $ref: '#/definitions/purchases.PurchaseHeader'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a purchase header
      tags:
      - purchases
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"